Basic Cube
Added a basic cube and some lightning as well as panning controll for mouse. Tried a lot of different integrations of the  three.js package but didn't manage to get any to work apart from the GitHub clone. Ignored the library for size, check documentation to re-clone it. If other option is to be pursued, go with cdn, not node package as it is not recommended by three.js